+++ libs/components/data-manager/projects/data-manager/src/modules/data-manager/data-manager-column-picker/data-manager-column-picker-impl.service.ts
@@ -1,0 +1,14 @@
+import {
+  Injectable
+} from '@angular/core';
+
+import {
+  SkyDataManagerColumnPickerComponent
+} from './data-manager-column-picker.component';
+
+@Injectable()
+export class SkyDataManagerColumnPickerImplService {
+  public getComponentType(): typeof SkyDataManagerColumnPickerComponent {
+    return SkyDataManagerColumnPickerComponent;
+  }
+}