+++ libs/components/data-manager/projects/data-manager/src/public-api.ts
@@ -1,0 +1,23 @@
+export * from './modules/data-manager/data-manager.module';
+export * from './modules/data-manager/data-manager.service';
+export * from './modules/data-manager/data-manager-filter-context';
+export * from './modules/data-manager/models/data-manager-column-picker-sort-strategy';
+export * from './modules/data-manager/models/data-manager-config';
+export * from './modules/data-manager/models/data-manager-filter-data';
+export * from './modules/data-manager/models/data-manager-init-args';
+export * from './modules/data-manager/models/data-manager-sort-option';
+export * from './modules/data-manager/models/data-manager-state';
+export * from './modules/data-manager/models/data-view-config';
+export * from './modules/data-manager/models/data-view-state';
+
+// Components and directives must be exported to support Angular's "partial" Ivy compiler.
+// Obscure names are used to indicate types are not part of public API.
+export { SkyDataManagerColumnPickerComponent as λ1 } from './modules/data-manager/data-manager-column-picker/data-manager-column-picker.component';
+export { SkyDataManagerComponent as λ2 } from './modules/data-manager/data-manager.component';
+export { SkyDataManagerToolbarLeftItemComponent as λ3 } from './modules/data-manager/data-manager-toolbar/data-manager-toolbar-left-item.component';
+export { SkyDataManagerToolbarPrimaryItemComponent as λ4 } from './modules/data-manager/data-manager-toolbar/data-manager-toolbar-primary-item.component';
+export { SkyDataManagerToolbarRightItemComponent as λ5 } from './modules/data-manager/data-manager-toolbar/data-manager-toolbar-right-item.component';
+export { SkyDataManagerToolbarSectionComponent as λ6 } from './modules/data-manager/data-manager-toolbar/data-manager-toolbar-section.component';
+export { SkyDataManagerToolbarComponent as λ7 } from './modules/data-manager/data-manager-toolbar/data-manager-toolbar.component';
+export { SkyDataViewComponent as λ8 } from './modules/data-manager/data-view.component';
+