+++ libs/components/data-manager/src/app/public/plugin-resources/data-manager-resources-provider.ts
@@ -1,8 +1,0 @@
-import {
-  SkyAppLocaleInfo,
-  SkyLibResourcesProvider
-} from '@skyux/i18n';
-
-export class SkyDataManagerResourcesProvider implements SkyLibResourcesProvider {
-  public getString: (localeInfo: SkyAppLocaleInfo, name: string) => string;
-}